LPS-189243 Extract common code to a base class
