+++ modules/apps/ai-creator/ai-creator-openai-web/src/main/java/com/liferay/ai/creator/openai/web/internal/display/context/AICreatorOpenAICompanyConfigurationDisplayContext.java
@@ -17,8 +17,6 @@ package com.liferay.ai.creator.openai.web.internal.display.context;
 import com.liferay.ai.creator.openai.configuration.manager.AICreatorOpenAIConfigurationManager;
 import com.liferay.portal.kernel.module.configuration.ConfigurationException;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 
 import javax.servlet.http.HttpServletRequest;
@@ -26,47 +24,36 @@ import javax.servlet.http.HttpServletRequest;
 /**
  * @author Lourdes Fernández Besada
  */
-public class AICreatorOpenAICompanyConfigurationDisplayContext {
+public class AICreatorOpenAICompanyConfigurationDisplayContext
+	extends BaseAICreatorOpenAIConfigurationDisplayContext {
 
 	public AICreatorOpenAICompanyConfigurationDisplayContext(
 		AICreatorOpenAIConfigurationManager aiCreatorOpenAIConfigurationManager,
 		HttpServletRequest httpServletRequest) {
 
+		super(httpServletRequest);
+
 		_aiCreatorOpenAIConfigurationManager =
 			aiCreatorOpenAIConfigurationManager;
-		_httpServletRequest = httpServletRequest;
 
 		_themeDisplay = (ThemeDisplay)httpServletRequest.getAttribute(
 			WebKeys.THEME_DISPLAY);
 	}
 
-	public String getAPIKey() throws ConfigurationException {
-		String apiKey = ParamUtil.getString(
-			_httpServletRequest, "apiKey", null);
-
-		if (apiKey != null) {
-			return apiKey;
-		}
-
+	@Override
+	protected String getAICreatorOpenAIAPIKey() throws ConfigurationException {
 		return _aiCreatorOpenAIConfigurationManager.
 			getAICreatorOpenAICompanyAPIKey(_themeDisplay.getCompanyId());
 	}
 
-	public boolean isEnabled() throws ConfigurationException {
-		String enabled = ParamUtil.getString(
-			_httpServletRequest, "enableOpenAI", null);
-
-		if (enabled != null) {
-			return GetterUtil.getBoolean(enabled);
-		}
-
+	@Override
+	protected boolean isAICreatorOpenAIEnabled() throws ConfigurationException {
 		return _aiCreatorOpenAIConfigurationManager.
 			isAICreatorOpenAICompanyEnabled(_themeDisplay.getCompanyId());
 	}
 
 	private final AICreatorOpenAIConfigurationManager
 		_aiCreatorOpenAIConfigurationManager;
-	private final HttpServletRequest _httpServletRequest;
 	private final ThemeDisplay _themeDisplay;
 
 }
