+++ modules/apps/ai-creator/ai-creator-openai-web/src/main/java/com/liferay/ai/creator/openai/web/internal/display/context/AICreatorOpenAIGroupConfigurationDisplayContext.java
@@ -17,8 +17,6 @@ package com.liferay.ai.creator.openai.web.internal.display.context;
 import com.liferay.ai.creator.openai.configuration.manager.AICreatorOpenAIConfigurationManager;
 import com.liferay.portal.kernel.module.configuration.ConfigurationException;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 
 import javax.servlet.http.HttpServletRequest;
@@ -26,30 +24,20 @@ import javax.servlet.http.HttpServletRequest;
 /**
  * @author Lourdes Fernández Besada
  */
-public class AICreatorOpenAIGroupConfigurationDisplayContext {
+public class AICreatorOpenAIGroupConfigurationDisplayContext
+	extends BaseAICreatorOpenAIConfigurationDisplayContext {
 
 	public AICreatorOpenAIGroupConfigurationDisplayContext(
 		AICreatorOpenAIConfigurationManager aiCreatorOpenAIConfigurationManager,
 		HttpServletRequest httpServletRequest) {
 
+		super(httpServletRequest);
+
 		_aiCreatorOpenAIConfigurationManager =
 			aiCreatorOpenAIConfigurationManager;
-		_httpServletRequest = httpServletRequest;
 
 		_themeDisplay = (ThemeDisplay)httpServletRequest.getAttribute(
 			WebKeys.THEME_DISPLAY);
-	}
-
-	public String getAPIKey() throws ConfigurationException {
-		String apiKey = ParamUtil.getString(
-			_httpServletRequest, "apiKey", null);
-
-		if (apiKey != null) {
-			return apiKey;
-		}
-
-		return _aiCreatorOpenAIConfigurationManager.
-			getAICreatorOpenAIGroupAPIKey(_themeDisplay.getScopeGroupId());
 	}
 
 	public boolean isCompanyEnabled() throws ConfigurationException {
@@ -57,14 +45,14 @@ public class AICreatorOpenAIGroupConfigurationDisplayContext {
 			isAICreatorOpenAICompanyEnabled(_themeDisplay.getCompanyId());
 	}
 
-	public boolean isEnabled() throws ConfigurationException {
-		String enabled = ParamUtil.getString(
-			_httpServletRequest, "enableOpenAI", null);
+	@Override
+	protected String getAICreatorOpenAIAPIKey() throws ConfigurationException {
+		return _aiCreatorOpenAIConfigurationManager.
+			getAICreatorOpenAIGroupAPIKey(_themeDisplay.getScopeGroupId());
+	}
 
-		if (enabled != null) {
-			return GetterUtil.getBoolean(enabled);
-		}
-
+	@Override
+	protected boolean isAICreatorOpenAIEnabled() throws ConfigurationException {
 		return _aiCreatorOpenAIConfigurationManager.
 			isAICreatorOpenAIGroupEnabled(
 				_themeDisplay.getCompanyId(), _themeDisplay.getScopeGroupId());
@@ -72,7 +60,6 @@ public class AICreatorOpenAIGroupConfigurationDisplayContext {
 
 	private final AICreatorOpenAIConfigurationManager
 		_aiCreatorOpenAIConfigurationManager;
-	private final HttpServletRequest _httpServletRequest;
 	private final ThemeDisplay _themeDisplay;
 
 }
